雁币类型 排名 获取课程状态接口修复

+++ common/src/main/java/com/yanzi/common/entity/college/level/LevelInfo.java
@@ -12,7 +12,7 @@ public class LevelInfo {
     @JSONField(serialize = false)
     private long minExp;
     @JSONField(serialize = false)
-    private long coin;
+    private double coin;
     @JSONField(serialize = false)
     private int valid;
 
@@ -64,11 +64,11 @@ public class LevelInfo {
         this.courseId = courseId;
     }
 
-    public long getCoin() {
+    public double getCoin() {
         return coin;
     }
 
-    public void setCoin(long coin) {
+    public void setCoin(double coin) {
         this.coin = coin;
     }
 }
